refactor: o component Table foi dividido em duas tabelas TableServidores e TableEstagiarios para melhor modularidade
